tools/detector/requirements.txt
click
pyyaml
pillow